fed_prospector/requirements.txt
requests>=2.31.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
lxml>=5.0.0
apscheduler>=3.10.0
click>=8.1.0
tqdm>=4.66.0
ijson>=3.2.0
